export master & worker to crates, NOT WORKING, but still commitiing, tcp issue, TODO:port to `remoc`
