+++ tester/it_works/output
@@ -1,0 +1,1 @@
+It really does work!
