+++ tester/master_w_single_worker/Cargo.toml
@@ -8,3 +8,5 @@ edition = "2021"
 [dependencies]
 clap = { version = "3.0", features = ["derive"] }
 runner = { path = "../../runner" }
+master = { path = "../../master" }
+worker = { path = "../../worker" }
