+++ tester/master_w_single_worker/output
@@ -1,0 +1,1 @@
+It really does work!
