+++ tester/master_w_single_worker/src/main.rs
@@ -15,99 +15,9 @@ struct Args {
 }
 
 pub mod blast {
-    use super::*;
-    use std::{
-        net::{TcpListener, TcpStream}, 
-        thread, 
-        io::{Read, Write}
-    };
-
-    pub struct Master;
-     
-    impl Master {
-        pub fn run() {
-            let listener = TcpListener::bind(format!("127.0.0.1:{}", runner::MASTER_PORT))
-                .expect("failed to bind master port");
-
-            for stream in listener.incoming() {
-                match stream {
-                    Ok(stream) => {
-                        // thread::spawn(move|| {
-                            // connection succeeded
-                            Master::handle_worker(stream)
-                        // });
-                    }
-                    Err(e) => {
-                        println!("Error: {}", e);
-                    }
-                }
-                break;
-            }
-            
-            // update runner
-            match TcpStream::connect(format!("127.0.0.1:{}", runner::RUNNER_PORT)) {
-                Ok(mut stream) => {
-                    stream.write(b"Done").unwrap();
-                }
-                Err(e) => {
-                    println!("Errr connecting to runner: {}", e);
-                }
-            }
-            // close the socket server
-            // drop(listener);
-        }
-
-        fn handle_worker(mut stream: TcpStream) {
-            let mut data = [0 as u8; 1024]; // using 1024 byte buffer
-            let size = stream.read(&mut data)
-                .expect(&format!("An error occurred, terminating connection with {}", stream.peer_addr().unwrap()));
-            if &data[..size] == b"Can I start?" {
-                stream.write(b"You must!").unwrap();
-            } else {
-                return;
-            }
-            thread::sleep(std::time::Duration::from_millis(1000));
-            let size = stream.read(&mut data)
-                .expect(&format!("An error occurred, terminating connection with {}", stream.peer_addr().unwrap()));
-            if &data[..size] == b"Task finished" {
-            } else {
-            }
-        }
-
-    }
-    
-    pub struct Worker;
-    impl Worker {
-        pub fn run() {
-            match TcpStream::connect(format!("127.0.0.1:{}", runner::MASTER_PORT)) {
-                Ok(mut stream) => {
-                    stream.write(b"Can I start?").unwrap();
-
-                    let mut data = [0 as u8; 1024]; // 
-                    let size = stream.read(&mut data)
-                        .expect(&format!("Failed to receive response from master on Question: Can I start?"));
-                    if &data[..size] == b"You must!" {
-                    } else {
-                    }
-                }, 
-                Err(e) => {
-                    println!("failded to connect Error: {}", e);
-                },
-            }
-
-        } 
-
-        pub fn terminate() {
-            match TcpStream::connect(format!("127.0.0.1:{}", runner::MASTER_PORT)) {
-                Ok(mut stream) => {
-                    stream.write(b"Task finished").unwrap();
-                }, 
-                Err(e) => {
-                    println!("failded to connect & update master about termiantion and task finish : {}", e);
-                },
-            }
-        }
-    }
+    use master::Master;
+    use worker::Worker;
+    use super::{Args, Parser};
 
     pub fn initialize() {
         let args = Args::parse();
